refactor(llmobs): address remaining sampling review comments
- Drop the redundant #samplerRate field; use sampler.rate() for the rebuild check.
- Move formatKnuthRate to the shared src/util.js so the LLMObs tagger no longer
  imports priority_sampler just for a string formatter.
- Reword the sampleRate JSDoc in index.d.ts and index.d.v5.ts ("honored at
  ingestion time").
- Test rates that exercise the formatter: 1/3 (6-decimal cap) and 0.25
  (trailing-zero stripping).
