+++ index.d.v5.ts
@@ -4205,7 +4205,7 @@ declare namespace tracer {
       /**
        * The proportion of LLM Observability traces to sample, between `0` and `1` (inclusive).
        * The decision is computed once per trace, propagated across services, and recorded on every
-       * span; spans are always sent and the backend honors the decision. Defaults to `1`.
+       * span; spans are always sent and the decision is honored at ingestion time. Defaults to `1`.
        * @env DD_LLMOBS_SAMPLE_RATE
        * Programmatic configuration takes precedence over the environment variables listed above.
        */
