+++ packages/dd-trace/src/priority_sampler.js
@@ -19,6 +19,7 @@ const RateLimiter = require('./rate_limiter')
 const Sampler = require('./sampler')
 const { setSamplingRules } = require('./startup-log')
 const SamplingRule = require('./sampling_rule')
+const { formatKnuthRate } = require('./util')
 
 const {
   SAMPLING_MECHANISM_DEFAULT,
@@ -35,19 +36,6 @@ const {
 } = require('./constants')
 
 const DEFAULT_KEY = 'service:,env:'
-
-/**
- * Formats a sampling rate as a string with up to 6 decimal digits and no trailing zeros.
- *
- * @param {number} rate
- */
-function formatKnuthRate (rate) {
-  const string = Number(rate).toFixed(6)
-  for (let i = string.length - 1; i > 0; i--) {
-    if (string[i] === '0') continue
-    return string.slice(0, i + (string[i] === '.' ? 0 : 1))
-  }
-}
 
 const defaultSampler = new Sampler(AUTO_KEEP)
 
@@ -396,4 +384,3 @@ class PrioritySampler {
 }
 
 module.exports = PrioritySampler
-module.exports.formatKnuthRate = formatKnuthRate
